requirements.txt
pandas
requests
beautifulsoup4
selenium
webdriver-manager
python-whois
tldextract
scikit-learn
numpy
fastapi
uvicorn
pydantic
python-multipart
torch
torchvision
transformers
ultralytics
joblib
Pillow